pyproject.toml
[tool.poetry]
name = "camera-grpc-stream"
version = "0.1.0"
description = "A USB camera gRPC streaming application using OpenCV"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.8.0"
grpcio = "^1.56.0"
grpcio-tools = "^1.56.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
jinja2 = "^3.1.5"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
